Reusing ERC20 tests for ERC777.
