+++ test/drafts/ERC777/ERC777.test.js
@@ -12,6 +12,10 @@ const {
   shouldBehaveLikeERC777SendBurnWithSendHook,
 } = require('./ERC777.behavior');
 
+const {
+  shouldBehaveLikeERC20,
+} = require('../../token/ERC20/ERC20.behavior');
+
 const ERC777 = artifacts.require('ERC777Mock');
 const ERC777SenderRecipientMock = artifacts.require('ERC777SenderRecipientMock');
 
@@ -41,6 +45,8 @@ contract('ERC777', function ([
       beforeEach(async function () {
         this.token = await ERC777.new(holder, initialSupply, name, symbol, granularity, defaultOperators);
       });
+
+      shouldBehaveLikeERC20(initialSupply, holder, anyone, defaultOperatorA);
 
       it.skip('does not emit AuthorizedOperator events for default operators', async function () {
         expectEvent.not.inConstructor(this.token, 'AuthorizedOperator'); // This helper needs to be implemented
